🚚 move / noteStore 파일 정리, store에 notebox,note 디렉터리 생성
