+++ src/components/noteComponents/EmotionCategoryForm.tsx
@@ -1,12 +1,12 @@
 'use client';
 
 import { EMOTION_CATEGORIES } from '@/constants/openai/category';
-import { useNoteStore } from '@/store/noteStore';
+import { useNoteStore } from '@/store/note/noteStore';
 const EmotionCategoryForm = () => {
   const { selectedEmotions, toggleEmotion } = useNoteStore();
   return (
     <div className="p-4">
-      <p className="text-xl font-semibold mb-4">
+      <p className="text-xl font-semibold mb-4">   
         여러 감정이 들으셨을것 같네요
       </p>
       <div className="grid grid-cols-3 gap-4">
